Move `EmailImporter` to a separate folder
